Update README.md
I am unfamiliar with Renesas processors and neither is the company I work for.  I can't find any good documentation on how to set up hardware peripherals.  The sales rep tells me to just use the code generator but this creates several extra files and is useless as far as porting code over from one processor to another.  Any input on set up of the timers and oscillators would be greatly appreciated.